Завершена страница аналитики

+++ Himchistka.Services/DTO/Analitic/DTOChartAnalitic.cs
@@ -9,7 +9,7 @@ namespace Himchistka.Services.DTO.Analitic
     public class DTOChartAnalitic
     {
         public List<string> Labels {  get; set; }
-        public List<decimal> Spendings {  get; set; }
+        public List<double> Spendings {  get; set; }
         public List<decimal> Profits {  get; set; }
 
     }
